move rseqc-junction-saturation JS code to assets
