+++ multiqc/modules/rseqc/rseqc.py
@@ -4,6 +4,7 @@
 
 from collections import OrderedDict
 import logging
+import os
 
 from multiqc import config
 from multiqc.modules.base_module import BaseMultiqcModule
@@ -44,6 +45,9 @@ class MultiqcModule(BaseMultiqcModule):
                 'infer_experiment',
                 'bam_stat'
             ]
+
+        # Add self.js to be included in template
+        self.js = { 'assets/js/multiqc_rseqc.js' : os.path.join(os.path.dirname(__file__), 'assets', 'js', 'multiqc_rseqc.js') }
 
         # Call submodule functions
         for sm in rseqc_sections:
